+++ tests/test_schema.py
@@ -1120,8 +1120,22 @@ def test_meta_nested_exclude():
 
         class Meta:
             exclude = ('blubb.foo',)
-    sch = ParentSchema()
+
     data = dict(bla=1, bli=2, blubb=dict(foo=42, bar=24, baz=242))
+
+    sch = ParentSchema()
+    result = sch.dump(data)
+    assert 'bla' in result
+    assert 'blubb' in result
+    assert 'bli' in result
+    child = result['blubb']
+    assert 'foo' not in child
+    assert 'bar' in child
+    assert 'baz' in child
+
+    # Test fields with dot notations in Meta.exclude on multiple instantiations
+    # Regression test for https://github.com/marshmallow-code/marshmallow/issues/1212
+    sch = ParentSchema()
     result = sch.dump(data)
     assert 'bla' in result
     assert 'blubb' in result